Change all NULL to nullptr. Add the new C++11 override to all relevant virtual functions. Fixed compiler warnings, (only warnings left are the deprecation warnings about glut).

+++ VirtualH89/Src/ROM.cpp
@@ -67,7 +67,7 @@ ROM::getROM(const char* filename,
     file.read((char*) buf, fileSize);
     file.close();
 
-    rom = new ROM(fileSize);
+    rom = new ROM((int) fileSize);
     rom->setBaseAddress(addr);
     rom->initialize(buf, fileSize);
     return rom;
